tests/wpt/tests/css/css-shapes/shape-outside/values/shape-outside-gradient-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Lists: getComputedStyle().listStyleImage</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes/#shape-outside-property">
<meta name="assert" content="list-style-image computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('shape-outside', 'linear-gradient(45deg, red calc(25% + 10px), blue 80%)', 'linear-gradient(45deg, rgb(255, 0, 0) calc(25% + 10px), rgb(0, 0, 255) 80%)');
test_computed_value('shape-outside', 'linear-gradient(calc(90deg - 45deg), rgb(0, 128, 0), rgb(0, 0, 255)', 'linear-gradient(45deg, rgb(0, 128, 0), rgb(0, 0, 255))');

test_computed_value('shape-outside', 'radial-gradient(circle at calc(10px + 5%), red, blue)', 'radial-gradient(circle at calc(5% + 10px) 50%, rgb(255, 0, 0), rgb(0, 0, 255))');
test_computed_value('shape-outside', 'radial-gradient(ellipse calc(-0.5em + 10px) calc(0.5em + 10px) at 20px 30px, rgb(255, 0, 0), rgb(0, 0, 255))', 'radial-gradient(0px 30px at 20px 30px, rgb(255, 0, 0), rgb(0, 0, 255))');
test_computed_value('shape-outside', 'radial-gradient(black 0% 0.5em, 25%, green 30% 60%, calc(100% * 3 / 4), white calc(100% - 20%) 100%', 'radial-gradient(rgb(0, 0, 0) 0%, rgb(0, 0, 0) 20px, 25%, rgb(0, 128, 0) 30%, rgb(0, 128, 0) 60%, 75%, rgb(255, 255, 255) 80%, rgb(255, 255, 255) 100%)');

test_computed_value('shape-outside', 'conic-gradient(from 0.5turn at 20% 30%, red, blue calc(50% - 25%))', 'conic-gradient(from 180deg at 20% 30%, rgb(255, 0, 0), rgb(0, 0, 255) 25%)');
</script>
</body>
</html>
